+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 > 2017-06-28
 
 - **Bugfix:** Fix select() not select table header checkbox ([#84])
+- **Bugfix:** Page reload when clicking on submit button ([#85])
 
 [master]: https://github.com/yyssc/ssc-grid/compare/v0.69.9...master
 [#84]: https://github.com/yyssc/ssc-grid/issues/84
+[#85]: https://github.com/yyssc/ssc-grid/issues/85
 
 ## [v0.69.9]
 > 2017-06-27